v6.10

+++ service.sh
@@ -1,9 +1,12 @@
 MODPATH=${0%/*}
-API=`getprop ro.build.version.sdk`
 
 # log
-exec 2>$MODPATH/debug.log
+LOGFILE=$MODPATH/debug.log
+exec 2>$LOGFILE
 set -x
+
+# var
+API=`getprop ro.build.version.sdk`
 
 # properties
 resetprop ro.audio.ignore_effects false
@@ -29,7 +32,7 @@ else
 fi
 PID=`pidof $SERVER`
 if [ "$PID" ]; then
-  killall $SERVER
+  killall $SERVER android.hardware.audio@4.0-service-mediatek
 fi
 
 # wait
@@ -110,10 +113,10 @@ fi
 
 # function
 stop_log() {
-FILE=$MODPATH/debug.log
-SIZE=`du $FILE | sed "s|$FILE||g"`
+SIZE=`du $LOGFILE | sed "s|$LOGFILE||g"`
 if [ "$LOG" != stopped ] && [ "$SIZE" -gt 50 ]; then
   exec 2>/dev/null
+  set +x
   LOG=stopped
 fi
 }
@@ -127,15 +130,11 @@ sleep 15
 stop_log
 NEXTPID=`pidof $SERVER`
 if [ "`getprop init.svc.$SERVER`" != stopped ]; then
-  until [ "$PID" != "$NEXTPID" ]; do
-    check_audioserver
-  done
-  killall $PROC
-  check_audioserver
+  [ "$PID" != "$NEXTPID" ] && killall $PROC
 else
   start $SERVER
-  check_audioserver
 fi
+check_audioserver
 }
 
 # check
